backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
websockets==14.1
yfinance==0.2.50
pandas==2.2.3
numpy==2.2.1
ta==0.11.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiohttp==3.11.11
python-multipart==0.0.20
google-api-python-client==2.155.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
